fixed extreme corner case

+++ src/structs/lepton_decoder.rs
@@ -337,8 +337,9 @@ pub fn read_coefficient_block<const ALL_PRESENT: bool, R: Read>(
     let mut output = AlignedBlock::default();
 
     // these are used as predictors for the number of non-zero edge coefficients
-    let mut eob_x: u8 = 0;
-    let mut eob_y: u8 = 0;
+    // do math in 32 bits since this is faster on most platforms
+    let mut eob_x: u32 = 0;
+    let mut eob_y: u32 = 0;
 
     let mut num_non_zeros_7x7_remaining = num_non_zeros_7x7 as usize;
 
@@ -369,8 +370,8 @@ pub fn read_coefficient_block<const ALL_PRESENT: bool, R: Read>(
             if coef != 0 {
                 // here we calculate the furthest x and y coordinates that have non-zero coefficients
                 // which is later used as a predictor for the number of edge coefficients
-                let bx = coord & 7;
-                let by = coord >> 3;
+                let bx = u32::from(coord) & 7;
+                let by = u32::from(coord) >> 3;
 
                 debug_assert!(bx > 0 && by > 0, "this does the DC and the lower 7x7 AC");
 
@@ -401,8 +402,8 @@ pub fn read_coefficient_block<const ALL_PRESENT: bool, R: Read>(
         qt,
         pt,
         num_non_zeros_7x7,
-        eob_x,
-        eob_y,
+        eob_x as u8,
+        eob_y as u8,
     )?;
 
     // step 3, read the DC coefficient (0,0 of the block)
